feat(fe): FaceIcon 컴포넌트 구현
