+++ apps/web/src/app/sign-in/page.css.ts
@@ -1,4 +1,4 @@
-import { keyframes, style } from '@vanilla-extract/css';
+import { style } from '@vanilla-extract/css';
 
 import { fadeIn } from '@/styles/keyframes.css';
 
@@ -11,18 +11,4 @@ export const container = style({
   minHeight: '100vh',
 
   animation: `${fadeIn} 0.5s ease`,
-});
-
-const animation = keyframes({
-  from: {
-    transform: 'rotate(0deg)',
-  },
-
-  to: {
-    transform: 'rotate(30deg)',
-  },
-});
-
-export const icon = style({
-  animation: `${animation} 1.5s ease infinite alternate`,
 });
