+++ apps/web/src/shared/components/FaceIcon/style.css.ts
@@ -1,0 +1,15 @@
+import { keyframes, style } from '@vanilla-extract/css';
+
+const animation = keyframes({
+  from: {
+    transform: 'rotate(0deg)',
+  },
+
+  to: {
+    transform: 'rotate(30deg)',
+  },
+});
+
+export const icon = style({
+  animation: `${animation} 1.5s ease infinite alternate`,
+});
